Add files via upload
This dataset provides monthly precipitation data for China with a spatial resolution of 0.25° (approximately 25 km) and a temporal range from January 2001 to December 2018. The data is stored in MATLAB `.mat` format. It was generated for mainland China by applying Bayesian Model Averaging (BMA) to integrate the CPC precipitation dataset and the MERRA2 precipitation dataset. Validation using CGDPA precipitation data has confirmed the reliability of the dataset.

The geographical coverage of the dataset is mainland China. For storage efficiency, the data is stored in a `.mat` file, with precipitation values measured in millimeters (1 mm). The `.mat` data can be processed using MATLAB software, with the data being extracted using the `load` function as follows:

```matlab
load BMA2P_2001_2018.mat
```

The data is stored as a 3-dimensional array (144 × 256 × 216), where the dimensions represent rows, columns, and time, respectively. A time value of 1 corresponds to January 2001, 2 corresponds to February 2001, and so on.